I forgot to commit...
Well, basically I spend the last two days so much focused on the project that this commit will be full of new things on the app. Home Screen, Details Page and MVC architecture are now within project. Initial tests were made reading a json file but now our controller is handle the http requests we need. Next step is to create a new and better widget to the details page and putting search filters to work.